+++ HISTORY.rst
@@ -3,7 +3,7 @@
 Release History
 ===============
 
-0.26.1
+0.27.0
 +++++++++++++++
 
 **General updates**